fix(flx-rs): remove bad chars from STR when using flx-rs
When using the flx-rs scoring backend, we need to run given string
values through fussy-remove-bad-char-fn to match the behavior of using
the flx scoring backend.

Without cleaning up bad characters, flx-rs seems like it works, but a
lot of result candidates do not appear, yielding a much smaller result
set compared to using flx. This fixes that, so results for both match.